+++ R/get_cam_op.R
@@ -23,8 +23,6 @@
 #'   returned by [camtrapR::cameraOperation()](
 #'   https://jniedballa.github.io/camtrapR/reference/cameraOperation.html).
 #'   Default: `FALSE`.
-#' @param datapkg Deprecated.
-#'   Use `package` instead.
 #' @param ... filter predicates for filtering on deployments.
 #' @return A matrix.
 #'   Row names always indicate the station ID.
@@ -46,16 +44,13 @@
 get_cam_op <- function(package = NULL,
                        ...,
                        station_col = "locationName",
-                       use_prefix = FALSE,
-                       datapkg = lifecycle::deprecated()) {
+                       use_prefix = FALSE) {
   # check camera trap data package
-  check_package(package, datapkg, "get_cam_op")
-  if (is.null(package) & !is.name(datapkg)) {
-    package <- datapkg
-  }
+  check_package(package)
   
   # Check that station_col is a single string
   assertthat::assert_that(assertthat::is.string(station_col))
+  
   # Check that station_col is one of the columns in deployments
   assertthat::assert_that(
     station_col %in% names(deployments(package)),